// web/generated/schema.ts
export interface Event extends SanityDocument {
  _type: "event"

  /**
   * Title — `localeString`
   *
   *
   */
  title?: LocaleString

  /**
   * Slug — `localeSlug`
   *
   * Click Generate.
   */
  slug?: LocaleSlug

  /**
   * Date and time — `datetime`
   *
   *
   */
  date?: string

  /**
   * Recurring event — `boolean`
   *
   *
   */
  recurring?: boolean

  /**
   * Eventbrite link — `url`
   *
   * Leave blank if the event is not on Eventbrite.
   */
  briteLink?: string

  /**
   * Body — `localeRichText`
   *
   *
   */
  body?: LocaleRichText

  /**
   * Main image — `image`
   *
   * Images should be jpeg of 1440px along the longest edge, 500-600k is best.
   */
  mainImage?: {
    _type: "image"
    asset: SanityReference<SanityImageAsset>
    crop?: SanityImageCrop
    hotspot?: SanityImageHotspot
  }

  /**
   * SEO title — `localeString`
   *
   * Displayed on Facebook and Twitter shares (max 60 characters).
   */
  seoTitle?: LocaleString

  /**
   * SEO description — `localeString`
   *
   * Displayed on Facebook and Twitter shares (max 65 characters).
   */
  seoDescription?: LocaleString

  /**
   * SEO Image — `image`
   *
   * Ideal size 1200 x 630px (if not added main image will be used).
   */
  seoImage?: {
    _type: "image"
    asset: SanityReference<SanityImageAsset>
    crop?: SanityImageCrop
    hotspot?: SanityImageHotspot
  }
}

/**
 * Exhibition
 *
 *
 */
export interface Exhibition extends SanityDocument {
  _type: "exhibition"

  /**
   * Title — `localeString`
   *
   *
   */
  title?: LocaleString

  /**
   * Slug — `localeSlug`
   *
   *
   */
  slug?: LocaleSlug

  /**
   * Start date — `date`
   *
   *
   */
  dateStart?: string

  /**
   * End date — `date`
   *
   *
   */
  dateEnd?: string

  /**
   * Body — `localeRichText`
   *
   *
   */
  body?: LocaleRichText

  /**
   * Main image — `captionImage`
   *
   *
   */
  mainImage?: CaptionImage

  /**
   * SEO title — `localeString`
   *
   * Displayed on Facebook and Twitter shares (max 60 characters).
   */
  seoTitle?: LocaleString

  /**
   * SEO description — `localeString`
   *
   * Displayed on Facebook and Twitter shares (max 65 characters).
   */
  seoDescription?: LocaleString

  /**
   * SEO Image — `image`
   *
   * Ideal size 1200 x 630px (if not added main image will be used).
   */
  seoImage?: {
    _type: "image"
    asset: SanityReference<SanityImageAsset>
    crop?: SanityImageCrop
    hotspot?: SanityImageHotspot
  }
}
// ...
export interface Post extends SanityDocument {
  _type: "post"

  /**
   * Title — `localeString`
   *
   *
   */
  title?: LocaleString

  /**
   * Body — `localeRichText`
   *
   *
   */
  body?: LocaleRichText

  /**
   * Categories — `array`
   *
   *
   */
  categories?: Array<SanityKeyedReference<Category>>

  /**
   * Published on — `datetime`
   *
   *
   */
  publishedAt?: string

  /**
   * Slug — `localeSlug`
   *
   *
   */
  slug?: LocaleSlug

  /**
   * Image — `image`
   *
   *
   */
  image?: {
    _type: "image"
    asset: SanityReference<SanityImageAsset>
    crop?: SanityImageCrop
    hotspot?: SanityImageHotspot
  }
}
// ...
export interface Category extends SanityDocument {
  _type: "category"

  /**
   * Title — `localeString`
   *
   *
   */
  title?: LocaleString
}

/**
 * Discipline
 *
 *
 */
export interface Discipline extends SanityDocument {
  _type: "discipline"

  /**
   * Title — `localeString`
   *
   *
   */
  title?: LocaleString
}

export type BlockContent = Array<
  | SanityKeyed<SanityBlock>
  | SanityKeyed<{
      _type: "image"
      asset: SanityReference<SanityImageAsset>
      crop?: SanityImageCrop
      hotspot?: SanityImageHotspot
    }>
>

export type CaptionImage = {
  _type: "image"
  asset: SanityReference<SanityImageAsset>
  crop?: SanityImageCrop
  hotspot?: SanityImageHotspot

  /**
   * Caption — `string`
   *
   *
   */
  caption?: string

  /**
   * Attribution — `string`
   *
   *
   */
  Caption?: string
}

export type LocaleRichText = {
  _type: "localeRichText"
  /**
   * English — `blockContent`
   *
   *
   */
  en?: BlockContent

  /**
   * Welsh — `blockContent`
   *
   *
   */
  cy?: BlockContent
}

export type LocaleSlug = {
  _type: "localeSlug"
  /**
   * English — `slug`
   *
   *
   */
  en?: { _type: "en"; current: string }

  /**
   * Welsh — `slug`
   *
   *
   */
  cy?: { _type: "cy"; current: string }
}

export type LocaleString = {
  _type: "localeString"
  /**
   * English — `string`
   *
   *
   */
  en?: string

  /**
   * Welsh — `string`
   *
   *
   */
  cy?: string
}